Move RelativeTimestampGenerator to the util package
