+++ app/src/main/java/org/simple/clinic/recentpatientsview/RecentPatientItem.kt
@@ -12,7 +12,7 @@ import org.simple.clinic.patient.Gender
 import org.simple.clinic.patient.displayIconRes
 import org.simple.clinic.recentpatientsview.SeeAllItem.SeeAllItemViewHolder
 import org.simple.clinic.summary.GroupieItemWithUiEvents
-import org.simple.clinic.summary.RelativeTimestamp
+import org.simple.clinic.util.RelativeTimestamp
 import org.simple.clinic.widgets.UiEvent
 import java.util.UUID
