+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryScreenControllerTest.kt
@@ -53,8 +53,10 @@ import org.simple.clinic.summary.PatientSummaryScreenControllerTest.GoBackToScre
 import org.simple.clinic.summary.addphone.MissingPhoneReminderRepository
 import org.simple.clinic.util.Just
 import org.simple.clinic.util.None
+import org.simple.clinic.util.RelativeTimestampGenerator
 import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.util.TestUtcClock
+import org.simple.clinic.util.Today
 import org.simple.clinic.util.randomMedicalHistoryAnswer
 import org.simple.clinic.util.toOptional
 import org.simple.clinic.widgets.UiEvent
